Organize commands in subfolders
