+++ src/commands/fun/cat.ts
