+++ src/commands/fun/joke.ts
