+++ src/commands/fun/math.ts
