+++ src/commands/fun/monke.ts
