+++ src/commands/util/ping.ts
